Generate UMAP for the final prediction based on cell patterns

Generate UMAP for the final prediction based on cell patterns

Usage

PlotUMAP(predictMatrix, prediction_result, n_component = 30, seed = 123, ...)

Arguments

predictMatrix

a wide cell by pattern matrix generated from GenerateInput function

prediction_result

Prediction result from PredictCellType

n_component

Number of PCA components to use (Default: 30)

seed

A number for random seed (Default: 123)

...

Additional arguments passed to uwot::umap (e.g., n_neighbors, metric).

Value

A list of two ggplot2 UMAP object.