Smooth cell by pattern matrix to reduce noise

Smooth cell by pattern matrix to reduce noise

Usage

smooth_matrix(predictMatrix, KNeighbor = 5)

Arguments

predictMatrix

A wide cell by pattern matrix generated from GenerateInput function

KNeighbor

Number of knn neighbors to use for smoothing (Default: 5)

Value

A wide cell by pattern matrix after smoothing and knn graph